+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ayazmo/types
 
+## 0.6.0
+
 ## 0.5.1
 
 ### Patch Changes