+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ayazmo/utils
 
+## 0.6.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ayazmo/types@0.6.0
+
 ## 0.5.1
 
 ### Patch Changes